requirements.txt
pdfminer.six
docx2txt
numpy
nltk
